requirements.txt
requests>=2.32.3
urllib3>=2.2.1
customtkinter>=5.2.2
PyYAML>=6.0.1
requests_ratelimiter>=0.6.0
aiohttp>=3.9.5
aiohttp-retry>=2.8.3